wip: fix waku.config.ts
