+++ e2e/fixtures/broken-links/waku.config.ts
@@ -1,10 +1,10 @@
 import { defineConfig } from 'waku/config';
 
 export default defineConfig({
-  middleware: () => [
-    import('waku/middleware/context'),
-    import('./src/redirects.js'),
-    import('waku/middleware/dev-server'),
-    import('waku/middleware/handler'),
+  middleware: [
+    'waku/middleware/context',
+    './src/redirects.js',
+    'waku/middleware/dev-server',
+    'waku/middleware/handler',
   ],
 });
