+++ e2e/fixtures/ssr-catch-error/waku.config.ts
@@ -1,11 +1,11 @@
 import { defineConfig } from 'waku/config';
 
 export default defineConfig({
-  middleware: () => [
-    import('waku/middleware/context'),
-    import('waku/middleware/dev-server'),
-    import('./src/middleware/validator.js'),
-    import('waku/middleware/handler'),
+  middleware: [
+    'waku/middleware/context',
+    'waku/middleware/dev-server',
+    './src/middleware/validator.js',
+    'waku/middleware/handler',
   ],
   /**
    * Base path for HTTP requests to indicate RSC requests.
